# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Manager API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ServerAuthType =
  (typeof ServerAuthType)[keyof typeof ServerAuthType];

export const ServerAuthType = {
  password: "password",
  key: "key",
} as const;

export type ServerStatus = (typeof ServerStatus)[keyof typeof ServerStatus];

export const ServerStatus = {
  unknown: "unknown",
  connected: "connected",
  disconnected: "disconnected",
  error: "error",
} as const;

export interface Server {
  id: number;
  name: string;
  host: string;
  port: number;
  username: string;
  authType: ServerAuthType;
  status: ServerStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateServerBodyAuthType =
  (typeof CreateServerBodyAuthType)[keyof typeof CreateServerBodyAuthType];

export const CreateServerBodyAuthType = {
  password: "password",
  key: "key",
} as const;

export interface CreateServerBody {
  name: string;
  host: string;
  port?: number;
  username: string;
  authType: CreateServerBodyAuthType;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  privateKey?: string | null;
  /** @nullable */
  passphrase?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateServerBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  port?: number | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  authType?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  privateKey?: string | null;
  /** @nullable */
  passphrase?: string | null;
  /** @nullable */
  notes?: string | null;
}

/**
 * @nullable
 */
export type ConnectionResultServerInfo = {
  os?: string;
  uptime?: string;
  hostname?: string;
} | null;

export interface ConnectionResult {
  success: boolean;
  message: string;
  /** @nullable */
  serverInfo?: ConnectionResultServerInfo;
}

export type FileEntryType = (typeof FileEntryType)[keyof typeof FileEntryType];

export const FileEntryType = {
  file: "file",
  directory: "directory",
  symlink: "symlink",
  other: "other",
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  /** @nullable */
  size?: number | null;
  permissions: string;
  owner: string;
  group: string;
  modifiedAt: string;
}

export interface FileListResponse {
  path: string;
  entries: FileEntry[];
  /** @nullable */
  parentPath?: string | null;
}

export interface FileContent {
  path: string;
  content: string;
  size: number;
  isBinary: boolean;
}

export interface WriteFileBody {
  path: string;
  content: string;
}

export interface CreateDirectoryBody {
  path: string;
}

export interface RenameFileBody {
  oldPath: string;
  newPath: string;
}

export interface ExecCommandBody {
  command: string;
  /** @nullable */
  cwd?: string | null;
}

export interface ExecResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  command: string;
}

export interface ProcessEntry {
  pid: number;
  user: string;
  cpu: number;
  mem: number;
  command: string;
  status: string;
}

export interface ProcessListResponse {
  processes: ProcessEntry[];
  count: number;
}

export type KillProcessBodySignal =
  (typeof KillProcessBodySignal)[keyof typeof KillProcessBodySignal];

export const KillProcessBodySignal = {
  SIGTERM: "SIGTERM",
  SIGKILL: "SIGKILL",
  SIGINT: "SIGINT",
} as const;

export interface KillProcessBody {
  signal?: KillProcessBodySignal;
}

export interface CpuStats {
  usage: number;
  cores: number;
  model: string;
}

export interface MemoryStats {
  total: number;
  used: number;
  free: number;
  usagePercent: number;
}

export interface DiskStats {
  filesystem: string;
  size: string;
  used: string;
  available: string;
  usePercent: string;
  mountedOn: string;
}

export interface ServerStats {
  cpu: CpuStats;
  memory: MemoryStats;
  disk: DiskStats[];
  uptime: string;
  loadAverage: number[];
  os: string;
  hostname: string;
  kernel: string;
}

export type ListFilesParams = {
  path?: string;
};

export type ReadFileParams = {
  path: string;
};

export type DeleteFileParams = {
  path: string;
  recursive?: boolean;
};