# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VPS Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers/{id}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a directory on a server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "/"
      responses:
        "200":
          description: File listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/files/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read file contents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write content to a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteFileBody"
      responses:
        "200":
          description: File written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/files/delete:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or directory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: recursive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/files/mkdir:
    post:
      operationId: createDirectory
      tags: [files]
      summary: Create a directory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDirectoryBody"
      responses:
        "200":
          description: Directory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/files/rename:
    post:
      operationId: renameFile
      tags: [files]
      summary: Rename or move a file/directory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFileBody"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/exec:
    post:
      operationId: execCommand
      tags: [terminal]
      summary: Execute a command on the server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecCommandBody"
      responses:
        "200":
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/processes:
    get:
      operationId: listProcesses
      tags: [processes]
      summary: List running processes on the server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Process list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessListResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/processes/{pid}/kill:
    post:
      operationId: killProcess
      tags: [processes]
      summary: Kill a process by PID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pid
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KillProcessBody"
      responses:
        "200":
          description: Process killed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory, symlink, other]
        size:
          type: ["integer", "null"]
        permissions:
          type: string
        owner:
          type: string
        group:
          type: string
        modifiedAt:
          type: string
      required:
        - name
        - path
        - type
        - permissions
        - owner
        - group
        - modifiedAt
    FileListResponse:
      type: object
      properties:
        path:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/FileEntry"
        parentPath:
          type: ["string", "null"]
      required:
        - path
        - entries
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
        isBinary:
          type: boolean
      required:
        - path
        - content
        - size
        - isBinary
    WriteFileBody:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    CreateDirectoryBody:
      type: object
      properties:
        path:
          type: string
      required:
        - path
    RenameFileBody:
      type: object
      properties:
        oldPath:
          type: string
        newPath:
          type: string
      required:
        - oldPath
        - newPath
    ExecCommandBody:
      type: object
      properties:
        command:
          type: string
        cwd:
          type: ["string", "null"]
      required:
        - command
    ExecResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        command:
          type: string
      required:
        - stdout
        - stderr
        - exitCode
        - command
    ProcessEntry:
      type: object
      properties:
        pid:
          type: integer
        user:
          type: string
        cpu:
          type: number
        mem:
          type: number
        command:
          type: string
        status:
          type: string
      required:
        - pid
        - user
        - cpu
        - mem
        - command
        - status
    ProcessListResponse:
      type: object
      properties:
        processes:
          type: array
          items:
            $ref: "#/components/schemas/ProcessEntry"
        count:
          type: integer
      required:
        - processes
        - count
    KillProcessBody:
      type: object
      properties:
        signal:
          type: string
          enum: [SIGTERM, SIGKILL, SIGINT]
          default: SIGTERM
...