# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all VPS servers
 */
export const ListServersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  port: zod.number(),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  status: zod.enum(["unknown", "connected", "disconnected", "error"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListServersResponse = zod.array(ListServersResponseItem);

/**
 * @summary Add a new VPS server
 */
export const createServerBodyPortDefault = 22;

export const CreateServerBody = zod.object({
  name: zod.string(),
  host: zod.string(),
  port: zod.number().default(createServerBodyPortDefault),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  password: zod.string().nullish(),
  privateKey: zod.string().nullish(),
  passphrase: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a server by ID
 */
export const GetServerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetServerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  port: zod.number(),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  status: zod.enum(["unknown", "connected", "disconnected", "error"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update server details
 */
export const UpdateServerParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateServerBody = zod.object({
  name: zod.string().nullish(),
  host: zod.string().nullish(),
  port: zod.number().nullish(),
  username: zod.string().nullish(),
  authType: zod.string().nullish(),
  password: zod.string().nullish(),
  privateKey: zod.string().nullish(),
  passphrase: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateServerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  port: zod.number(),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  status: zod.enum(["unknown", "connected", "disconnected", "error"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a server
 */
export const DeleteServerParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Test SSH connection to server
 */
export const TestServerConnectionParams = zod.object({
  id: zod.coerce.number(),
});

export const TestServerConnectionResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  serverInfo: zod
    .object({
      os: zod.string().optional(),
      uptime: zod.string().optional(),
      hostname: zod.string().optional(),
    })
    .nullish(),
});

/**
 * @summary List files in a directory on a server
 */
export const ListFilesParams = zod.object({
  id: zod.coerce.number(),
});

export const listFilesQueryPathDefault = `/`;

export const ListFilesQueryParams = zod.object({
  path: zod.coerce.string().default(listFilesQueryPathDefault),
});

export const ListFilesResponse = zod.object({
  path: zod.string(),
  entries: zod.array(
    zod.object({
      name: zod.string(),
      path: zod.string(),
      type: zod.enum(["file", "directory", "symlink", "other"]),
      size: zod.number().nullish(),
      permissions: zod.string(),
      owner: zod.string(),
      group: zod.string(),
      modifiedAt: zod.string(),
    }),
  ),
  parentPath: zod.string().nullish(),
});

/**
 * @summary Read file contents
 */
export const ReadFileParams = zod.object({
  id: zod.coerce.number(),
});

export const ReadFileQueryParams = zod.object({
  path: zod.coerce.string(),
});

export const ReadFileResponse = zod.object({
  path: zod.string(),
  content: zod.string(),
  size: zod.number(),
  isBinary: zod.boolean(),
});

/**
 * @summary Write content to a file
 */
export const WriteFileParams = zod.object({
  id: zod.coerce.number(),
});

export const WriteFileBody = zod.object({
  path: zod.string(),
  content: zod.string(),
});

export const WriteFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Delete a file or directory
 */
export const DeleteFileParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteFileQueryParams = zod.object({
  path: zod.coerce.string(),
  recursive: zod.coerce.boolean().optional(),
});

export const DeleteFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Create a directory
 */
export const CreateDirectoryParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateDirectoryBody = zod.object({
  path: zod.string(),
});

export const CreateDirectoryResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Rename or move a file/directory
 */
export const RenameFileParams = zod.object({
  id: zod.coerce.number(),
});

export const RenameFileBody = zod.object({
  oldPath: zod.string(),
  newPath: zod.string(),
});

export const RenameFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Execute a command on the server
 */
export const ExecCommandParams = zod.object({
  id: zod.coerce.number(),
});

export const ExecCommandBody = zod.object({
  command: zod.string(),
  cwd: zod.string().nullish(),
});

export const ExecCommandResponse = zod.object({
  stdout: zod.string(),
  stderr: zod.string(),
  exitCode: zod.number(),
  command: zod.string(),
});

/**
 * @summary List running processes on the server
 */
export const ListProcessesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListProcessesResponse = zod.object({
  processes: zod.array(
    zod.object({
      pid: zod.number(),
      user: zod.string(),
      cpu: zod.number(),
      mem: zod.number(),
      command: zod.string(),
      status: zod.string(),
    }),
  ),
  count: zod.number(),
});

/**
 * @summary Kill a process by PID
 */
export const KillProcessParams = zod.object({
  id: zod.coerce.number(),
  pid: zod.coerce.number(),
});

export const killProcessBodySignalDefault = `SIGTERM`;

export const KillProcessBody = zod.object({
  signal: zod
    .enum(["SIGTERM", "SIGKILL", "SIGINT"])
    .default(killProcessBodySignalDefault),
});

export const KillProcessResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get system resource statistics for a server
 */
export const GetServerStatsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetServerStatsResponse = zod.object({
  cpu: zod.object({
    usage: zod.number(),
    cores: zod.number(),
    model: zod.string(),
  }),
  memory: zod.object({
    total: zod.number(),
    used: zod.number(),
    free: zod.number(),
    usagePercent: zod.number(),
  }),
  disk: zod.array(
    zod.object({
      filesystem: zod.string(),
      size: zod.string(),
      used: zod.string(),
      available: zod.string(),
      usePercent: zod.string(),
      mountedOn: zod.string(),
    }),
  ),
  uptime: zod.string(),
  loadAverage: zod.array(zod.number()),
  os: zod.string(),
  hostname: zod.string(),
  kernel: zod.string(),
});