# lib/api-zod/src/generated/types/fileContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileContent {
  path: string;
  content: string;
  size: number;
  isBinary: boolean;
}