# lib/api-zod/src/generated/types/fileListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntry } from "./fileEntry";

export interface FileListResponse {
  path: string;
  entries: FileEntry[];
  /** @nullable */
  parentPath?: string | null;
}