# lib/api-zod/src/generated/types/renameFileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RenameFileBody {
  oldPath: string;
  newPath: string;
}